+++ src/commands/run-extension.js
@@ -1,10 +1,49 @@
 const vscode = require("vscode");
 const Mustache = require("mustache");
+const camelCase = require("lodash.camelcase");
+const snakeCase = require("lodash.snakecase");
+const kebabCase = require("lodash.kebabcase");
+const upperCase = require("lodash.uppercase");
+const lowerCase = require("lodash.lowercase");
+const startCase = require("lodash.startcase");
 const {
   DEFAULT_TEMPLATES_FOLDER_PATH,
   DEFAULT_CONFIG_FILE_NAME,
 } = require("../_constants");
 const { getDirStructure, buildDirStructure } = require("../utils");
+
+const stringFunctions = {
+  toCamelCase: function () {
+    return function (text, render) {
+      return camelCase(render(text));
+    };
+  },
+  toSnakeCase: function () {
+    return function (text, render) {
+      return snakeCase(render(text));
+    };
+  },
+  toKebabCase: function () {
+    return function (text, render) {
+      return kebabCase(render(text));
+    };
+  },
+  toUpperCase: function () {
+    return function (text, render) {
+      return upperCase(render(text));
+    };
+  },
+  toLowerCase: function () {
+    return function (text, render) {
+      return lowerCase(render(text));
+    };
+  },
+  toPascalCase: function () {
+    return function (text, render) {
+      return startCase(camelCase(render(text))).replace(/ /g, "");
+    };
+  },
+};
 
 const runExtension = (folderURI) => {
   const defaultTemplatesFolderURI = vscode.Uri.file(
@@ -40,11 +79,16 @@ const runExtension = (folderURI) => {
       }
     }
 
+    const contentView = {
+      ...promptResults,
+      ...stringFunctions,
+    };
+
     const structure = await getDirStructure({
       path: folderURI.path,
       dirURI: templateURI,
       onNameCopy: (name) => Mustache.render(name, promptResults),
-      onContentCopy: (content) => Mustache.render(content, promptResults),
+      onContentCopy: (content) => Mustache.render(content, contentView),
     });
 
     buildDirStructure(structure);